requirements.txt
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
SQLAlchemy==1.2.1
flask-testing-0.7.1
blinker-1.4